Remove surface_tint_color from Material controls and themes
Eliminates the surface_tint_color property from various Material controls, styles, and theme classes in the Flet SDK. It is not recommended to use in Flutter: "This is not recommended for use. Material 3 spec introduced a set of tone-based surfaces and surface containers in its ColorScheme, which provide more flexibility. The intention is to eventually remove surface tint color from the framework."